requirements.txt
flask
flask-socketio
psutil
python-dotenv
